warnings: zero-warning release prep — fix all rustc/clippy diagnostics
Real fixes (no behavior change):
- eBPF: drop duplicate inline(always), remove unnecessary unsafe on
  XDP_TCP_CT.insert, snat_port via let-else, drop unused aya-log-ebpf dep
- smoltcp-edge: declare the socket-tcp-timestamp feature the code cfgs on
- collapse nested ifs into let-chains, struct-update syntax for
  Default-derived values, Iterator::contains/checked_div/is_multiple_of/
  repeat_n/c-string literal/Error::other idioms
- rename h3_probe -> h3-probe (kebab-case bin name)
- extract type aliases for complex types (EifelCheckpoint,
  KernelSnapshotProvider, PendingPermit)
- add is_empty() next to len() in transport sched types

Scoped allows (justified):
- #[allow(clippy::too_many_arguments)] on 6 internal fns where
  signature refactors would be churn without gain
- vendored libinjectionrs: allow(style, complexity) at crate root —
  upstream code is not restyled; correctness/suspicious stay enabled
- allow(unused_mut/unused_variables) for linux-gated params on non-Linux

Verified on Linux (Debian 12, rustc 1.98.1 + pinned nightly eBPF):
cargo check --all-targets and cargo clippy --all-targets
-D clippy::correctness -D clippy::suspicious emit zero warnings.

+++ crates/cloud-node-transport/src/accecn.rs
@@ -322,7 +322,7 @@ impl AccEcn {
     /// Feed the peer's ACE field (3 bits) from an incoming ACK — sender
     /// side. Returns the new CE *packet* delta (wrap-safe).
     pub fn on_ack_ace(&mut self, ace: u8) -> u64 {
-        let delta = ((ace as u64 + ACE_MOD - self.peer_ace as u64) % ACE_MOD) as u64;
+        let delta = (ace as u64 + ACE_MOD - self.peer_ace as u64) % ACE_MOD;
         self.peer_ace = ace;
         self.peer_cep_packets += delta;
         delta
